added seperate tracer.cpp/hpp file for exec history tracing
