+++ Makefile
@@ -3,10 +3,10 @@ LIBS = -lcapstone -lreadline
 TESTLIB = gtest
 FLAGS = -Wall -std=c++20
 
-OBJECTS = build/main.o build/dbg.o build/fmt.o build/dbgtypes.o build/elf.o build/elftypes.o build/disass.o
+OBJECTS = build/main.o build/dbg.o build/tracer.o build/fmt.o build/dbgtypes.o build/elf.o build/elftypes.o build/disass.o
 ELF_OBJECTS = build/elf.o build/elftypes.o build/disass.o build/fmt.o build/dbgtypes.o
 DISASS_OBJECTS = build/disass.o build/fmt.o
-DBG_OBJECTS = build/dbg.o build/dbgtypes.o build/elf.o build/elftypes.o build/disass.o  build/fmt.o
+DBG_OBJECTS = build/dbg.o build/tracer.o build/dbgtypes.o build/elf.o build/elftypes.o build/disass.o  build/fmt.o
 
 all: $(OBJECTS)
 	$(CC) $(OBJECTS) -o wg  $(LIBS) $(FLAGS)
